# Conflicts:
#	core/data/src/main/java/com/android/mediproject/core/data/remote/di/RepositoryModule.kt
#	core/domain/src/main/java/com/android/mediproject/core/domain/sign/GetUserUseCase.kt
#	core/model/src/main/java/com/android/mediproject/core/model/user/remote/UserResponse.kt
#	core/network/src/main/java/com/android/mediproject/core/network/module/AwsNetwork.kt
#	feature/mypage/src/main/java/com/android/mediproject/feature/mypage/mypagemore/MyPageMoreDialogFragment.kt
#	feature/penalties/src/main/java/com/android/mediproject/feature/penalties/recentpenaltylist/PenaltyListAdapter.kt

+++ core/model/src/main/java/com/android/mediproject/core/model/requestparameters/DeleteCommentParameter.kt
@@ -12,6 +12,6 @@ import kotlinx.serialization.Serializable
  */
 @Serializable
 data class DeleteCommentParameter(
-    @SerialName("commentId") val commentId: Int, // 1
-    @SerialName("medicineId") val medicineId: Int // 41
+    @SerialName("commentId") val commentId: Long, // 1
+    @SerialName("medicineId") val medicineId: Long
 )
